qcow2 boot: prepare a cloud-init image per VM
Create a cloud-init image and register it as a disk. This allow to drop
the big "if cloud-init image" from the XML template.

Drop from the test the big list of system call. This is getting a bit
complicated. We need a better/cleaver way to check we do the right
calls.

The network is still not properly configured.